==> plagh_icu/3.data extract_labs&vitals.ipynb <==
%config IPCompleter.greedy=True
import numpy as np
import pandas as pd

import matplotlib.pyplot as plt
import time
import re
import scipy
from datetime import time, timedelta
import seaborn as sns
pd.set_option('display.max_columns', 100)

from dateutil.parser import parse

import warnings
warnings.filterwarnings("ignore")

from pandas.api.types import CategoricalDtype

def trans_time(t):
    if t == 'None':
        return np.datetime64('NaT')
    elif pd.isna(t):
        return np.datetime64('NaT')
    else:
        return np.datetime64(t)

def trans_float(df):
    for i in df.columns:
        if df[i].dtype == object:
            if ((df[i]== 'None')|(df[i]== 'Null')).any():
                df[i].replace('None',np.nan,inplace = True)
                df[i].replace('Null',np.nan,inplace = True)
                df[i] = df[i].astype(float)
            else:
                df[i] = df[i].astype(float)

def trans_time(t):
    if t == 'None':
        return np.datetime64('NaT')
    elif pd.isna(t):
        return np.datetime64('NaT')
    else:
        return parse(t)

def read_largefile(path, encoding):
    test = pd.read_csv(path, sep=',',encoding = encoding,nrows= 10000)
    nan_count_list = []
    col_sel_list = []
    for i in range(len(test.columns.to_list())):
        nan_count_list.append((pd.isna(test.iloc[:,i]).sum()/10000)*100)

    for i in range(len(nan_count_list)):
        if nan_count_list[i]<=50:
            col_sel_list.append(test.columns.to_list()[i])
    
    reader = pd.read_csv(path,usecols= col_sel_list, sep=',', iterator=True,encoding = encoding)
    loop = True
    chunkSize = 100000
    chunks = []
    while loop:
        try:
            chunk = reader.get_chunk(chunkSize)
            chunks.append(chunk)
        except StopIteration:
            loop = False
            print("Iteration is stopped.")
    df = pd.concat(chunks, ignore_index=True)
    return df

# 生命体征

vitals = read_largefile('../Local_hosp_data/vitals.csv', encoding = 'gb2312')

vitals.VITAL_SIGNS.value_counts()

vitals = vitals[vitals.VITAL_SIGNS.isin(['脉搏','呼吸','血压high','血压Low','血氧','体温','血糖','心率','尿量','总入量','总出量','呼吸机辅助'])].\
drop_duplicates().dropna(axis = 0,subset = ['VITAL_SIGNS_VALUES'])

#vitals.to_csv('vitals.csv', index = False)

# 实验室检查

lab = read_largefile('../Local_hosp_data/lab.csv', encoding = 'gb18030')

['ALP', 'ALT', 'APPT', 'AST','BNP', 'CRP', 'D_dimer', 'IL_6', 'INR', 'LYN', 'PA', 'PCT', 'PT', 'PTT',
'albumin', 'amylase', 'bands', 'basophil', 'bilirubin',
'bilirubin_direct', 'ca', 'chloride', 'ck', 'ck_mb', 'creatinine',
'eosinophil', 'fibrinogen', 'gamma_GT', 'glucose', 'hct_pcv',
'hemoglobin', 'lipase', 'mb', 'mch', 'mchc', 'mcv', 'mg', 'monocytes',
'mpv', 'p', 'platelet', 'potassium', 'rbc', 'rdw', 'sodium', 'urea','wbc']

#mimic 
# mg,50960,mg/dl; mmol/L
# phosphate,50970,mg/dl; mmol/L
# 'rbc', 51279, m/uL; 10^12/L
#  'platelet',51265, K/uL; 10^9/L
#  'rdw',51277,%; %
#  'chloride',50902,mEq/L; mmol/L
#  'sodium',50983,mEq/L; mmol/L
#  'hemoglobin',51222,g/dL; g/L

lab[(lab['LABEL'] == 'p')].UNITS.value_counts()

# icustay_detail

final_cohort = pd.read_csv('final_cohort.csv')

final_cohort = final_cohort[final_cohort['LOS_ICU']>0]

for i in ['ADMISSION_DATE_TIME','DISCHARGE_DATE_TIME','TF_ADMISSION_DATE_TIME', 'TF_DISCHARGE_DATE_TIME']:
    final_cohort[i] = final_cohort[i].apply(trans_time)

# 实验室检查值

lab_sel = lab[lab['PATIENT_ID'].isin(final_cohort['PATIENT_ID'].to_list())]

lab_bc = pd.merge(final_cohort[['PATIENT_ID','TF_ADMISSION_DATE_TIME']],lab_sel,how = 'left', on = ['PATIENT_ID'])

lab_bc[(lab_bc['LABEL'] == 'wbc')].PRINT_CONTEXT.value_counts()

lab_bc.info()

for i in ['REQUESTED_DATE_TIME','RESULTS_RPT_DATE_TIME']:
    lab_bc[i] = lab_bc[i].apply(trans_time)

lab_bc_1stday = lab_bc[((lab_bc['REQUESTED_DATE_TIME'] - lab_bc['TF_ADMISSION_DATE_TIME']).dt.total_seconds()/3600<=48)&\
                       ((lab_bc['REQUESTED_DATE_TIME'] - lab_bc['TF_ADMISSION_DATE_TIME']).dt.total_seconds()/3600>=-24)]

lab_bc_1stday.dropna(axis = 0, subset = ['RESULT'], inplace = True)

lab_bc_1stday = lab_bc_1stday.drop_duplicates()

def sub_f(x):
    if x == 'nan':
        return np.nan
    else:
        x_togroup = re.search("\d+(\.\d+)?",x)
        if x_togroup == None:
            return np.nan
        else:
            return x_togroup.group()


lab_bc_1stday['RESULT'] = lab_bc_1stday['RESULT'].astype(str).apply(sub_f).astype(float)

lab_bc_1stday[(lab_bc_1stday['LABEL'] == 'APPT')].PRINT_CONTEXT.value_counts()

def blood_def_clean(df):
    if df['LABEL'] in ['bands','LYN','basophil','monocytes','eosinophil'] and (df['RESULT']>=1 or df['RESULT']<0):
        return np.nan    
    else:
        return df['RESULT']

lab_bc_1stday['RESULT'] = lab_bc_1stday.apply(blood_def_clean,axis = 1)

lab_count = lab_bc_1stday.LABEL.value_counts()

lab_bc_sel = lab_bc_1stday[lab_bc_1stday.LABEL.isin(lab_count.head(47).index.to_list())]

lab_bc_sel = lab_bc_sel[['PATIENT_ID', 'TF_ADMISSION_DATE_TIME', 
       'REQUESTED_DATE_TIME',  'LABEL','RESULT']].drop_duplicates().dropna()

lab_bc_sel_1st = lab_bc_sel.groupby(['PATIENT_ID','TF_ADMISSION_DATE_TIME','LABEL']).apply(lambda x: x.sort_values(['REQUESTED_DATE_TIME'],ascending = True).head(1))

lab_1st = lab_bc_sel_1st.reset_index(drop = True)

lab_1st.drop(columns = ['REQUESTED_DATE_TIME'],inplace = True)

#lab_1st[lab_1st['PATIENT_ID'] =='S118067120'].set_index(['PATIENT_ID','TF_ADMISSION_DATE_TIME','LABEL']).unstack()

lab_1st = lab_1st.set_index(['PATIENT_ID','TF_ADMISSION_DATE_TIME','LABEL']).unstack()

lab_name = []
for i in lab_1st.reset_index().columns.to_list()[2:]:
    lab_name.append(i[1])

lab_name.insert(0,lab_1st.reset_index().columns.to_list()[0][0])
lab_name.insert(1,lab_1st.reset_index().columns.to_list()[1][0])

lab_1st_re = lab_1st.reset_index()

lab_1st_re.columns = lab_name

#lab_1st_re[lab_1st_re['C-反应蛋白测定'].str.contains(r'[^((0-9)|\.)]')]

lab_1st_re_re = lab_1st_re.copy()

lab_1st_re_re.columns

lab_ = pd.merge(final_cohort[['PATIENT_ID','TF_ADMISSION_DATE_TIME']],lab_1st_re_re, how = 'left', on = ['PATIENT_ID','TF_ADMISSION_DATE_TIME'])

lab_ = lab_.drop_duplicates()

def lookup_miss(df):
    nadf = pd.DataFrame(columns=['nan_nums','nan_ratios'])
    for i in df.columns:
        nadf.loc[i,'nan_nums'] = pd.isna(df[i]).sum()
        nadf.loc[i,'nan_ratios'] = (nadf.loc[i,'nan_nums']/df.shape[0])*100
    return nadf.sort_values(by = 'nan_ratios', ascending = False)

lookup_miss(lab_)

# 生命体征

vitals.VITAL_SIGNS.value_counts()[0:20]

vitals = vitals[vitals.VITAL_SIGNS.isin(['脉搏','呼吸','血压high','血压Low','血氧','体温','血糖','心率'])].\
drop_duplicates().dropna(axis = 0,subset = ['VITAL_SIGNS_VALUES'])

vitals_bc = pd.merge(final_cohort[['PATIENT_ID','VISIT_ID','TF_ADMISSION_DATE_TIME']],vitals,how = 'left', on = ['PATIENT_ID','VISIT_ID'])

vitals_bc['TIME_POINT'] = vitals_bc['TIME_POINT'].apply(trans_time)

vitals_bc_1stday = vitals_bc[((vitals_bc['TIME_POINT'] - vitals_bc['TF_ADMISSION_DATE_TIME']).dt.total_seconds()/3600>=-24)&\
((vitals_bc['TIME_POINT'] - vitals_bc['TF_ADMISSION_DATE_TIME']).dt.total_seconds()/3600<=24)]

vitals_bc_1stday.columns

vitals_bc_1stday = vitals_bc_1stday[['PATIENT_ID', 'TF_ADMISSION_DATE_TIME', 'TIME_POINT', 'VITAL_SIGNS', 'VITAL_SIGNS_VALUES']]

vitals_bc_1stday_1st = vitals_bc_1stday.groupby(['PATIENT_ID','TF_ADMISSION_DATE_TIME', 'VITAL_SIGNS']).apply(lambda x: x.sort_values(['TIME_POINT'],ascending = True).head(1))

vitals_bc_1stday_1st = vitals_bc_1stday_1st.reset_index(drop = True)

vitals_1st = vitals_bc_1stday_1st.drop(columns = ['TIME_POINT']).set_index(['PATIENT_ID','TF_ADMISSION_DATE_TIME','VITAL_SIGNS']).unstack()

vitals_name = []
for i in vitals_1st.reset_index().columns.to_list()[2:]:
    vitals_name.append(i[1])

vitals_name.insert(0,vitals_1st.reset_index().columns.to_list()[0][0])
vitals_name.insert(1,vitals_1st.reset_index().columns.to_list()[1][0])

vitals_1st = vitals_1st.reset_index()

vitals_1st.columns = vitals_name

pd.isna(vitals_1st).sum()

vitals_1st.drop(columns = ['心率','血糖'], inplace = True)

vitals_1st.columns

vitals_1st.columns = ['PATIENT_ID', 'TF_ADMISSION_DATE_TIME', 'Temperature', 'RR', 'HR', 'DBP','SBP', 'SPO2']

bc_cohort = pd.merge(final_cohort,lab_,how = 'left', on = ['PATIENT_ID', 'TF_ADMISSION_DATE_TIME']).\
merge(vitals_1st, how = 'left', on = ['PATIENT_ID', 'TF_ADMISSION_DATE_TIME'])

bc_cohort['GENDER'] = bc_cohort['GENDER'].apply(lambda x: 1 if x == 'M' else 0)

bc_cohort = pd.get_dummies(data = bc_cohort, columns= ['ICU_TYPE'])

bc_cohort = bc_cohort.drop(columns = ['previous_mdrpos_lessthan7d','previous_mdrneg_lessthan7d','los_icu_exc','stay_chart_included'])

icustay_detail = pd.read_csv('../Local_hosp_data/icustay_detail.csv',encoding = 'gb18030')
icustay_detail = icustay_detail[~pd.isna(icustay_detail['ICU_TYPE'])]
icustay_detail['TF_ADMISSION_DATE_TIME'] = icustay_detail['TF_ADMISSION_DATE_TIME'].apply(trans_time)

icustay_detail['icustay_rank']= icustay_detail.groupby(['PATIENT_ID'])['TF_ADMISSION_DATE_TIME'].rank()

icustay_detail['icustay_rank'].value_counts()

bc_cohort.drop(columns = ['icustay_rank'], inplace = True)

bc_cohort = bc_cohort.merge(icustay_detail[['PATIENT_ID', 'TF_ADMISSION_DATE_TIME','icustay_rank']], how = 'left', on = ['PATIENT_ID', 'TF_ADMISSION_DATE_TIME'])

#住院次数
bc_cohort['hospstay_seq'] = bc_cohort['VISIT_ID']

#住院时长
bc_cohort['los_hospital'] = (bc_cohort['DISCHARGE_DATE_TIME'] - bc_cohort['ADMISSION_DATE_TIME']).dt.total_seconds()/(3600*24)

#死亡信息
mort = pd.read_csv('../Local_hosp_data/mort.csv')

mort['DOD_HOSP'] = mort['DOD_HOSP'].apply(trans_time)

bc_cohort = pd.merge(bc_cohort, mort, how = 'left', on = ['PATIENT_ID','VISIT_ID'])

bc_cohort['DOD'] = \
bc_cohort[['ADMISSION_DATE_TIME','DISCHARGE_DATE_TIME','DOD_HOSP']].\
apply(lambda x: 1 if ((x[2]-x[0]).total_seconds()>=0) and ((x[2]-x[1]).total_seconds()<=0) else 0, axis = 1)

bc_cohort = bc_cohort.drop_duplicates()

#bc_cohort_icu = bc_cohort[(bc_cohort['ICU_TYPE_ICU'] == 1)|(bc_cohort['ICU_TYPE_EICU'] == 1)]#.drop_duplicates()

bc_cohort_icu = bc_cohort[(bc_cohort['ICU_TYPE_ICU'] == 1)]

icustay_detail.columns

hw = icustay_detail[['PATIENT_ID', 'VISIT_ID','BODY_HEIGHT','BODY_WEIGHT']]
hw['BODY_HEIGHT'] = hw['BODY_HEIGHT'].astype(float)
hw['BODY_WEIGHT'] = hw['BODY_WEIGHT'].astype(float)
hw['BMI'] = hw['BODY_WEIGHT']/(hw['BODY_HEIGHT']/100)**2

bc_cohort_icu = pd.merge(bc_cohort_icu,hw, how = 'left', on = ['PATIENT_ID', 'VISIT_ID'])

bc_cohort_icu.hospstay_seq

bc_cohort_icu[bc_cohort_icu['AGE']>=18][['PATIENT_ID','TF_ADMISSION_DATE_TIME']].drop_duplicates().shape

bc_cohort_icu['BODY_HEIGHT'] = bc_cohort_icu['BODY_HEIGHT'].apply(lambda x:np.nan if ((x<=100) or (x>=250)) else x)
bc_cohort_icu['BODY_WEIGHT'] = bc_cohort_icu['BODY_WEIGHT'].apply(lambda x:np.nan if ((x<=30) or (x>=230)) else x)
bc_cohort_icu['BMI'] = bc_cohort_icu['BODY_WEIGHT']/(bc_cohort_icu['BODY_HEIGHT']/100)**2

bc_cohort_icu = bc_cohort_icu.drop_duplicates()

#bc_cohort_icu = pd.read_csv('mdro_cohort.csv', encoding = 'ANSI')

bc_cohort_icu.stay_chart_included_mdr.value_counts()

pd.isna(bc_cohort_icu).sum().sort_values(ascending =False).head(25)

bc_cohort['PCT_nan'] = bc_cohort['PCT'].apply(lambda x:1 if pd.isna(x) else 0)

bc_cohort[bc_cohort['ICU_TYPE_ICU']==1].stay_chart_included_mdr.value_counts()

missing_percent = lab_.isnull().mean(axis=1)
lab_filtered = lab_[missing_percent < 0.2]
lab_filtered = lab_filtered[~pd.isna(lab_filtered['wbc'])]

bc_cohort_icu = bc_cohort_icu.merge(lab_filtered[['PATIENT_ID','TF_ADMISSION_DATE_TIME']],how = 'inner', on = ['PATIENT_ID','TF_ADMISSION_DATE_TIME'] )

bc_cohort_icu.shape

bc_cohort_icu.drop_duplicates().to_csv('mdro_cohort.csv', index  = False, encoding = 'utf-8')

# 经R语言插补缺失值

bc_cohort_median_fill = pd.read_csv('MDRO_afterimp_Rmice.csv',encoding = 'utf-8')

bc_cohort_median_fill['monocytes_abs'] = bc_cohort_median_fill['monocytes'] * bc_cohort_median_fill['wbc']
bc_cohort_median_fill['neutrophils_abs'] = bc_cohort_median_fill['bands'] * bc_cohort_median_fill['wbc']
bc_cohort_median_fill['eosinophil_abs'] = bc_cohort_median_fill['eosinophil'] * bc_cohort_median_fill['wbc']
bc_cohort_median_fill['basophil_abs'] = bc_cohort_median_fill['basophil'] * bc_cohort_median_fill['wbc']
bc_cohort_median_fill['lymphocytes_abs'] = bc_cohort_median_fill['LYN'] * bc_cohort_median_fill['wbc']
bc_cohort_median_fill['lymphocytes_abs'] = bc_cohort_median_fill['lymphocytes_abs'].apply(lambda x:0.0001 if x == 0 else x)

bc_cohort_median_fill['RPR'] = bc_cohort_median_fill['rdw']/bc_cohort_median_fill['platelet']
bc_cohort_median_fill['NLR'] = bc_cohort_median_fill['neutrophils_abs']/bc_cohort_median_fill['lymphocytes_abs']

bc_cohort_median_fill['SBP'] = bc_cohort_median_fill['SBP'].apply(lambda x:0.0001 if x == 0 else x)
bc_cohort_median_fill['shock_index'] = bc_cohort_median_fill['HR']/bc_cohort_median_fill['SBP']

bc_cohort_median_fill.icustay_rank.value_counts()

bc_cohort_median_fill.TF_ADMISSION_DATE_TIME = bc_cohort_median_fill.TF_ADMISSION_DATE_TIME.apply(trans_time)

bc_cohort_median_fill.to_csv('plagh_totableone.csv',index = False)

bc_cohort_median_fill['year'] = bc_cohort_median_fill.TF_ADMISSION_DATE_TIME.dt.year

pd.pivot_table(bc_cohort_median_fill, index = ['year'], values = ['stay_chart_included_mdr'], aggfunc= ['count','mean'])

train_df = bc_cohort_median_fill[bc_cohort_median_fill['year'].isin([2008,2009,2010,2011,2012,2013,2014,2015,2016])]
val_df = bc_cohort_median_fill[bc_cohort_median_fill['year'].isin([2017,2018,2019])]

train_df.to_csv('train_df.csv',index = False)
val_df.to_csv('val_df.csv',index = False)
