==> plagh_icu/5.select parameters.ipynb <==
# Packages

import warnings
warnings.filterwarnings("ignore")

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
%matplotlib inline
pd.set_option('max_columns',200)

import matplotlib as mpl
mpl.style.use('ggplot')
sns.set_style('white')

import sklearn
from sklearn.linear_model import LogisticRegression
from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import RandomForestClassifier
from xgboost.sklearn import XGBClassifier
from sklearn.svm import SVC
from sklearn.neighbors import KNeighborsClassifier

import sys
import os
import tempfile
from sklearn.preprocessing import StandardScaler

from sklearn.model_selection import GridSearchCV
from sklearn.model_selection import train_test_split
from sklearn.model_selection import ShuffleSplit
from sklearn.model_selection import cross_val_score
from sklearn.model_selection import StratifiedKFold

from sklearn import metrics
from sklearn.metrics import confusion_matrix
from sklearn.metrics import plot_roc_curve
from sklearn.metrics import roc_auc_score
from sklearn.metrics import roc_curve,auc
from sklearn.metrics import accuracy_score

# Read  the data

dev = pd.read_csv('train_df.csv')
val = pd.read_csv('val_df.csv')

dev.columns

features = [
'los_hosp_beforeicu',
 'ALP',
 'BNP',
 'CRP',
 'INR',
 'PCT',
 'IL_6',
 'bilirubin',
 'ck',
 'fibrinogen',
 'rbc',
 'sodium',
 'urea',
 'HR',
 'SBP'
]
target = ['stay_chart_included_mdr']

x_train = dev[features]
y_train = dev[target]

x_extval = val[features]
y_extval = val[target]

y_extval.value_counts()

train_tot = dev[features + target]
extval_tot = val[features + target]

def lookup_miss(df):
    nadf = pd.DataFrame(columns=['nan_nums','nan_ratios'])
    for i in df.columns:
        nadf.loc[i,'nan_nums'] = pd.isna(df[i]).sum()
        nadf.loc[i,'nan_ratios'] = (nadf.loc[i,'nan_nums']/df.shape[0])*100
    return nadf.sort_values(by = 'nan_ratios', ascending = False)

lookup_miss(x_train)

scaler = StandardScaler()

x_train_scaler = scaler.fit_transform(x_train)
x_extval_scaler = scaler.transform(x_extval)

skf = StratifiedKFold(n_splits=5, random_state=None, shuffle=False)

## Logistic Regression

lr_model = LogisticRegression(random_state=2021)
         

param_dict1 = {'penalty':['l1','l2'],
             'C':[0.00001,0.0001,0.001,0.01,0.1,1,10,100]
             }
estimator_grid_lr = GridSearchCV(lr_model,param_grid=param_dict1,cv=skf,n_jobs = -1,scoring = 'roc_auc')

estimator_grid_lr.fit(x_train_scaler,y_train)

print("----------------------------------------")
print('最佳参数：\n',estimator_grid_lr.best_params_)
print('最佳结果：\n',estimator_grid_lr.best_score_)
print('最佳估计器：\n',estimator_grid_lr.best_estimator_)

roc_auc_score(y_extval, estimator_grid_lr.predict_proba(x_extval_scaler)[:,1])

## Random Forest

forest_model1 = RandomForestClassifier(random_state=2021)
         

param_dict1 = {'n_estimators':[100,150,200,250,300],
             'max_depth':[3,4,5,6,7],
            'min_samples_leaf':[7,9,11,12,13,15]
             }
estimator_grid_forest1 = GridSearchCV(forest_model1,param_grid=param_dict1,cv=skf,n_jobs = -1,scoring = 'roc_auc')

estimator_grid_forest1.fit(x_train,y_train)

print("----------------------------------------")
print('最佳参数：\n',estimator_grid_forest1.best_params_)
print('最佳结果：\n',estimator_grid_forest1.best_score_)
print('最佳估计器：\n',estimator_grid_forest1.best_estimator_)

roc_auc_score(y_extval, estimator_grid_forest1.predict_proba(x_extval)[:,1])

forest_model1 = RandomForestClassifier(random_state=2021)
         

param_dict1 = {'n_estimators':[110,120,130,140,150,160,170,180,190,200],
             'max_depth':[5,6,7],
            'min_samples_leaf':[14,15,16,17,18,19]
             }
estimator_grid_forest1 = GridSearchCV(forest_model1,param_grid=param_dict1,cv=skf,n_jobs = 5,scoring = 'roc_auc')

estimator_grid_forest1.fit(x_train,y_train)

print("----------------------------------------")
print('最佳参数：\n',estimator_grid_forest1.best_params_)
print('最佳结果：\n',estimator_grid_forest1.best_score_)
print('最佳估计器：\n',estimator_grid_forest1.best_estimator_)

roc_auc_score(y_extval, estimator_grid_forest1.predict_proba(x_extval)[:,1])

forest_model1 = RandomForestClassifier(random_state=2021)
         

param_dict1 = {'n_estimators':[50,60,80,90,100,110,120],
             'max_depth':[5,6,7,8],
            'min_samples_leaf':[12,13,14,15]
             }
estimator_grid_forest1 = GridSearchCV(forest_model1,param_grid=param_dict1,cv=skf,n_jobs = 5,scoring = 'roc_auc')

estimator_grid_forest1.fit(x_train,y_train)

print("----------------------------------------")
print('最佳参数：\n',estimator_grid_forest1.best_params_)
print('最佳结果：\n',estimator_grid_forest1.best_score_)
print('最佳估计器：\n',estimator_grid_forest1.best_estimator_)

roc_auc_score(y_extval, estimator_grid_forest1.predict_proba(x_extval)[:,1])

## XGBOOST

xgboost_model1_1 = XGBClassifier(random_state = 2021)
         

param_dict1_1 = {
              'n_estimators':[15,20,25,35,45,50,70,80],
            'max_depth':[2,3,4,5,6],
            'min_child_weight':[1,2,4,6],
            'learning_rate':[0.001,0.01,0.1]
             }
estimator_grid_xgb1_1 = GridSearchCV(xgboost_model1_1,param_grid=param_dict1_1,cv=skf, n_jobs = -1,scoring = 'roc_auc',verbose = 20)

estimator_grid_xgb1_1.fit(x_train,y_train)

print("----------------------------------------")
print('最佳参数：\n',estimator_grid_xgb1_1.best_params_)
print('最佳结果：\n',estimator_grid_xgb1_1.best_score_)
print('最佳估计器：\n',estimator_grid_xgb1_1.best_estimator_)

roc_auc_score(y_extval, estimator_grid_xgb1_1.predict_proba(x_extval)[:,1])

xgboost_model1_1 = XGBClassifier(random_state = 2021)
         

param_dict1_1 = {
              'n_estimators':[40,41,42,43,44,45,46,47,48,49,50,51,52,53,54,55],
            'max_depth':[2,3,4,5],
            'min_child_weight':[2,3,4,5],
     'learning_rate':[0.01,0.1]
             }
estimator_grid_xgb1_1 = GridSearchCV(xgboost_model1_1,param_grid=param_dict1_1,cv=skf, n_jobs = 10,scoring = 'roc_auc',verbose = 20)

estimator_grid_xgb1_1.fit(x_train,y_train)

print("----------------------------------------")
print('最佳参数：\n',estimator_grid_xgb1_1.best_params_)
print('最佳结果：\n',estimator_grid_xgb1_1.best_score_)
print('最佳估计器：\n',estimator_grid_xgb1_1.best_estimator_)

roc_auc_score(y_extval, estimator_grid_xgb1_1.predict_proba(x_extval)[:,1])

## SVM

svm_model = SVC(kernel='linear', probability = True, random_state = 2021)
         

param_dict1 = {
             'C':[1e-06, 0.00001,0.0001,0.001,0.01,0.1,1,10,100]
             }
estimator_grid_svm = GridSearchCV(svm_model,param_grid=param_dict1,cv=skf,n_jobs = 5,scoring = 'roc_auc')

estimator_grid_svm.fit(x_train_scaler,y_train)

print("----------------------------------------")
print('最佳参数：\n',estimator_grid_svm.best_params_)
print('最佳结果：\n',estimator_grid_svm.best_score_)
print('最佳估计器：\n',estimator_grid_svm.best_estimator_)

roc_auc_score(y_extval, estimator_grid_svm.predict_proba(x_extval_scaler)[:,1]

## KNN

knn_model = KNeighborsClassifier()
         

param_dict1 = {
             'n_neighbors':[5,20,35,40,45,50,60,70,80,100]
             }
estimator_grid_knn = GridSearchCV(knn_model,param_grid=param_dict1,cv=skf,n_jobs = 5,scoring = 'roc_auc')

estimator_grid_knn.fit(x_train_scaler,y_train)

print("----------------------------------------")
print('最佳参数：\n',estimator_grid_knn.best_params_)
print('最佳结果：\n',estimator_grid_knn.best_score_)
print('最佳估计器：\n',estimator_grid_knn.best_estimator_)

roc_auc_score(y_extval, estimator_grid_knn.predict_proba(x_extval_scaler)[:,1])

knn_model = KNeighborsClassifier()
         

param_dict2 = {
             'n_neighbors':[46,47,48,49,50,51,52,53,54,55,56,57,58,59]
             }
estimator_grid_knn2 = GridSearchCV(knn_model,param_grid=param_dict2,cv=skf,n_jobs = 5,scoring = 'roc_auc')

estimator_grid_knn2.fit(x_train_scaler,y_train)

print("----------------------------------------")
print('最佳参数：\n',estimator_grid_knn2.best_params_)
print('最佳结果：\n',estimator_grid_knn2.best_score_)
print('最佳估计器：\n',estimator_grid_knn2.best_estimator_)

roc_auc_score(y_extval, estimator_grid_knn2.predict_proba(x_extval_scaler)[:,1])
